FEATURE: moves Filters interface to options types files
