+++ libs/spirit-islander/shared/util/src/lib/functions/get-boards.ts
@@ -1,6 +1,5 @@
 import { BOARDS } from '../data';
-import { BalancedBoardName, Board } from '../types';
-import { Filters } from './filters.interface';
+import { BalancedBoardName, Board, Filters } from '../types';
 import { getOptions } from './get-options';
 
 export function getBoards(filters: Filters<BalancedBoardName> = {}) {
