+++ libs/spirit-islander/shared/util/src/lib/functions/get-maps.ts
@@ -1,6 +1,5 @@
 import { MAPS } from '../data';
-import { Map, MapName } from '../types';
-import { Filters } from './filters.interface';
+import { Filters, Map, MapName } from '../types';
 import { getOptions } from './get-options';
 
 export function getMaps(filters: Filters<MapName> = {}) {
