+++ libs/spirit-islander/shared/util/src/lib/types/option.ts
@@ -12,3 +12,8 @@ export interface DifficultyOption<TName extends string> extends Option<TName> {
 export interface ExpansionOption<TName extends string> extends Option<TName> {
   expansion?: Expansion;
 }
+
+export interface Filters<TName extends string> {
+  names?: readonly TName[];
+  expansions?: readonly Expansion[];
+}
